[MEG] meg master into v3.6

+++ console/views/team.py
@@ -1,10 +1,11 @@
 # -*- coding: utf8 -*-
 import logging
 
+from django.conf import settings
 from django.core.paginator import Paginator, PageNotAnInteger, EmptyPage
 from django.db.models import Q
 from rest_framework.response import Response
-
+import re
 from backends.services.exceptions import *
 from backends.services.resultservice import *
 from console.services.enterprise_services import enterprise_services
@@ -14,12 +15,10 @@ from console.services.perm_services import perm_services
 from console.services.region_services import region_services
 from console.views.base import JWTAuthApiView
 from www.models import Tenants
-from console.repositories.plugin import config_group_repo, config_item_repo
 from www.perms import PermActions, get_highest_identity
-from www.utils.crypt import make_uuid
 from www.utils.return_message import general_message, error_message
-from console.services.plugin import plugin_service, plugin_version_service
 from console.repositories.perm_repo import role_repo
+from console.repositories.region_repo import region_repo
 
 logger = logging.getLogger("default")
 
@@ -253,9 +252,9 @@ class TeamUserView(JWTAuthApiView):
                 for identity in perms_identitys_list:
                     if identity == "access":
                         role_info_list.append({"role_name": identity, "role_id": None})
-                    role_id = role_repo.get_role_id_by_role_name(identity)
-                    role_info_list.append({"role_name": identity, "role_id": role_id})
-
+                    else:
+                        role_id = role_repo.get_role_id_by_role_name(identity)
+                        role_info_list.append({"role_name": identity, "role_id": role_id})
                 for role in perms_role_list:
                     role_name = role_repo.get_role_name_by_role_id(role)
                     role_info_list.append({"role_name": role_name, "role_id": role})
@@ -524,6 +523,10 @@ class TeamDelView(JWTAuthApiView):
             if service_count >= 1:
                 result = general_message(400, "failed", "当前团队内有应用,不可以删除")
                 return Response(result, status=400)
+            tenants = team_services.get_current_user_tenants(self.user.user_id)
+            if len(tenants) == 1:
+                return Response(general_message(409, "you have to keep one team","您必须保留一个团队"),status=409)
+
             status = team_services.delete_tenant(tenant_name=team_name)
             if not status:
                 result = general_message(code, "delete a tenant successfully", "删除团队成功")
@@ -673,3 +676,74 @@ class TeamDetailView(JWTAuthApiView):
             logger.exception(e)
             result = error_message(e.message)
             return Response(result, status=result["code"])
+
+
+class TeamRegionInitView(JWTAuthApiView):
+    def post(self, request):
+        """
+        初始化团队和数据中心信息
+        ---
+        parameters:
+            - name: team_alias
+              description: 团队别名
+              required: true
+              type: string
+              paramType: form
+            - name: region_name
+              description: 数据中心名称
+              required: true
+              type: string
+              paramType: form
+        """
+        try:
+            team_alias = request.data.get("team_alias", None)
+            region_name = request.data.get("region_name", None)
+            if not team_alias:
+                return Response(general_message(400, "team alias is null", "团队名称不能为空"), status=400)
+            if not region_name:
+                return Response(general_message(400, "region name is null", "请选择数据中心"), status=400)
+            r = re.compile(u'^[a-zA-Z0-9_\\-\u4e00-\u9fa5]+$')
+            if not r.match(team_alias.decode("utf-8")):
+                return Response(general_message(400, "team alias is not allow", "组名称只支持中英文下划线和中划线"), status=400)
+            team = team_services.get_team_by_team_alias(team_alias)
+            if team:
+                return Response(general_message(409,"region alias is exist","团队名称{0}已存在".format(team_alias)),status=409)
+            region = region_repo.get_region_by_region_name(region_name)
+            if not region:
+                return Response(general_message(404, "region not exist", "需要开通的数据中心{0}不存在".format(region_name)), status=404)
+            enterprise = enterprise_services.get_enterprise_by_enterprise_id(self.user.enterprise_id)
+            if not enterprise:
+                return Response(general_message(404, "user's enterprise is not found","无法找到用户所在的数据中心"))
+
+            code, msg, team = team_services.create_team(self.user, enterprise, [region_name], team_alias)
+            if not code:
+                return Response(general_message(code, "create team error", msg), status=code)
+            perm_info = {
+                "user_id": self.user.user_id,
+                "tenant_id": team.ID,
+                "identity": "owner",
+                "enterprise_id": enterprise.pk
+            }
+            perm_services.add_user_tenant_perm(perm_info)
+            # 创建用户在企业的权限
+            user_services.make_user_as_admin_for_enterprise(self.user.user_id, enterprise.enterprise_id)
+            # 为团队开通默认数据中心并在数据中心创建租户
+            code, msg, tenant_region = region_services.create_tenant_on_region(team.tenant_name, team.region)
+            if code != 200:
+                return Response(general_message(code, "create tenant on region error", msg), status=code)
+            # 公有云，如果没有领过资源包，为开通的数据中心领取免费资源包
+            if settings.MODULES.get('SSO_LOGIN'):
+                result = region_services.get_enterprise_free_resource(tenant_region.tenant_id, enterprise.enterprise_id,
+                                                                      tenant_region.region_name, self.user.nick_name)
+                logger.debug("get free resource on [{}] to team {}: {}".format(tenant_region.region_name,
+                                                                               team.tenant_name,
+                                                                               result))
+            self.user.is_active = True
+            self.user.save()
+
+            result = general_message(200, "success", "初始化成功")
+
+        except Exception as e:
+            logger.exception(e)
+            result = error_message(e.message)
+        return Response(result, status=result["code"])
